--- MVP/Team #3/templates/index.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Linear Regression with Gradient Descent</title>
-</head>
-<body>
-    <h1>Linear Regression Model</h1>
-
-    <!-- Error message -->
-    {% if error_message %}
-        <div style="color: red;">
-            <p>{{ error_message }}</p>
-        </div>
-    {% endif %}
-
-    <!-- Form to upload dataset and input parameters -->
-    <form method="POST" enctype="multipart/form-data">
-        <label for="dataset">Upload dataset (CSV):</label>
-        <input type="file" name="dataset" accept=".csv"><br>
-
-        <label for="learning_rate">Learning Rate:</label>
-        <input type="text" id="learning_rate" name="learning_rate" value="{{ learning_rate }}" required><br>
-
-        <label for="iterations">Iterations:</label>
-        <input type="text" id="iterations" name="iterations" value="{{ iterations }}" required><br>
-
-        <label for="regularization_type">Regularization Type:</label>
-        <select name="regularization_type" required>
-            <option value="None" {% if regularization_type == "None" %}selected{% endif %}>None</option>
-            <option value="Lasso" {% if regularization_type == "Lasso" %}selected{% endif %}>Lasso</option>
-            <option value="Ridge" {% if regularization_type == "Ridge" %}selected{% endif %}>Ridge</option>
-        </select><br>
-
-        <label for="regularization_param">Regularization Parameter:</label>
-        <input type="text" name="regularization_param" value="{{ regularization_param }}" required><br>
-
-        <label for="cost_function">Cost Function:</label>
-        <select name="cost_function" required>
-            <option value="MSE" {% if cost_function == "MSE" %}selected{% endif %}>Mean Squared Error (MSE)</option>
-            <option value="MAE" {% if cost_function == "MAE" %}selected{% endif %}>Mean Absolute Error (MAE)</option>
-            <option value="R2" {% if cost_function == "R2" %}selected{% endif %}>R-squared (R²)</option>
-        </select><br>
-
-        <!-- Dynamic column selection -->
-        {% if columns %}
-            <label for="x_column">Select Feature Column (X):</label>
-            <select name="x_column" required>
-                {% for column in columns %}
-                    <option value="{{ column }}">{{ column }}</option>
-                {% endfor %}
-            </select><br>
-
-            <label for="y_column">Select Target Column (Y):</label>
-            <select name="y_column" required>
-                {% for column in columns %}
-                    <option value="{{ column }}">{{ column }}</option>
-                {% endfor %}
-            </select><br>
-        {% endif %}
-
-        <input type="submit" value="Train Model">
-    </form>
-
-    <!-- Display results -->
-    {% if plot_path %}
-        <h2>Results:</h2>
-        <p>Slope (m): {{ m }}</p>
-        <p>Intercept (b): {{ b }}</p>
-        <p>Cost Function: {{ cost_function }}</p>
-
-        <h3>Cost History:</h3>
-        <ul>
-            {% for cost in cost_history %}
-                <li>{{ cost }}</li>
-            {% endfor %}
-        </ul>
-
-        <h2>Regression Plot:</h2>
-        <img src="{{ url_for('static', filename='years_salary_linear_regression_plot.png') }}" alt="Regression Plot">
-    {% endif %}
-</body>
-</html>
