+++ MVP/Team#2/Reg/main.py
