+++ MVP/Team#3/static/years_salary_linear_regression_plot.png
