+++ MVP/Team#8/django_demo/Dockerfile
