+++ MVP/Team#8/django_demo/entrypoint.sh
