+++ MVP/Team#8/django_demo/manage.py
