+++ MVP/Team#8/django_demo/regression/views.py
